+++ doc/CHANGES.rst
@@ -8,13 +8,13 @@ http://docs.zope.org/zope2/releases/.
 2.12.26 (unreleased)
 --------------------
 
+- LP #930812: Scrub headers a bit more.
 
 2.12.25 (2012-09-18)
 --------------------
 
 - Explicitly declared ZTUtils APIs as public (repairs breakages in apps
   following fix for LP #1047318).
-
 
 2.12.24 (2012-09-09)
 --------------------